Better max client body size

+++ incs/Location.hpp
@@ -6,7 +6,7 @@
 /*   By: hlevi <hlevi@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/08 14:55:36 by hlevi             #+#    #+#             */
-/*   Updated: 2023/02/24 12:21:30 by hlevi            ###   ########.fr       */
+/*   Updated: 2023/03/13 12:45:47 by hlevi            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,6 +26,6 @@ struct	Location
 	std::string					root;
     std::string					auto_index;
     std::string					cgi_dir;
-	int							max_client_body_size;
+	unsigned long int			max_client_body_size;
 };
 }
